# eims-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExamFlow EIMS API
  description: |
    Exam Information Management System API for UP Tacloban.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requests per 15 minutes per IP
    - Rate limit headers included in responses
    
    ## WebSocket
    Real-time updates available at `ws://localhost:3001/ws?token=<access_token>`
  version: 1.0.0
  contact:
    name: ExamFlow Support
    email: support@examflow.up.edu.ph
  license:
    name: Educational Use
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://examflow.up.edu.ph/api/v1
    description: Production server
...
paths:
...
  /schedules:
    get:
      tags: [Schedules]
      summary: List schedules
      security:
        - bearerAuth: []
      parameters:
        - name: examId
          in: query
          schema:
            type: string
            format: uuid
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of schedules
    post:
      tags: [Schedules]
      summary: Create schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Schedule created
...
  /students:
    get:
      tags: [Students]
      summary: List students
      description: Get paginated list of students (Faculty/Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: search
          in: query
          schema:
            type: string
        - name: program
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of students
  /students/sync:
    post:
      tags: [Students]
      summary: Sync students from parent app
      description: Sync student data from parent application API (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                students:
                  type: array
                  items:
                    $ref: '#/components/schemas/StudentSyncData'
      responses:
        '200':
          description: Sync completed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    CreateScheduleRequest:
      type: object
      required: [examId, startTime, endTime]
      properties:
        examId:
          type: string
          format: uuid
        section:
          type: string
        room:
          type: string
        meetLink:
          type: string
          format: uri
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
...
    StudentSyncData:
      type: object
      required: [email, firstName, lastName, studentNumber, program, yearLevel]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        studentNumber:
          type: string
        program:
          type: string
        yearLevel:
          type: integer
        section:
          type: string
...